folder refactor: src/ ftw. Reorganizing lighthouse core's source
