+++ src/aggregator.js
