+++ src/aggregators/aggregate.js
