+++ src/aggregators/is-accessible/index.js
