+++ src/aggregators/is-performant/index.js
