+++ src/aggregators/is-secure/index.js
