+++ src/aggregators/is-sized-for-mobile-screen/index.js
