+++ src/aggregators/launches-with-splash-screen/index.js
