+++ src/aggregators/will-get-add-to-homescreen-prompt/index.js
