+++ src/auditor.js
