+++ src/audits/manifest/icons-min-192.js
