+++ src/audits/mobile-friendly/display.js
