+++ src/gatherers/manifest.js
