+++ src/gatherers/service-worker.js
