+++ src/gatherers/theme-color.js
