+++ src/helpers/emulation.js
