+++ src/helpers/frame-load-recorder.js
