+++ src/helpers/manifest-parser.js
