+++ src/helpers/traces/devtools-timeline-model.js
